requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4
    # via
    #   -c requirements.txt
    #   black
astroid==2.5.3
    # via pylint
attrs==20.3.0
    # via
    #   -c requirements.txt
    #   pytest
autoflake==1.4
    # via -r requirements-dev.in
bandit==1.7.0
    # via -r requirements-dev.in
black==20.8b1
    # via -r requirements-dev.in
certifi==2020.12.5
    # via
    #   -c requirements.txt
    #   requests
chardet==4.0.0
    # via
    #   -c requirements.txt
    #   requests
click==7.1.2
    # via
    #   -c requirements.txt
    #   black
coverage-badge==1.0.1
    # via -r requirements-dev.in
coverage==5.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
deprecation==2.1.0
    # via testcontainers
docker==5.0.0
    # via testcontainers
flake8==3.9.1
    # via -r requirements-dev.in
freezegun==1.1.0
    # via -r requirements-dev.in
gitdb==4.0.7
    # via gitpython
gitpython==3.1.15
    # via bandit
greenlet==1.0.0
    # via sqlalchemy
idna==2.10
    # via
    #   -c requirements.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via
    #   -r requirements-dev.in
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r requirements-dev.in
packaging==20.9
    # via
    #   deprecation
    #   pytest
pathspec==0.8.1
    # via black
pbr==5.5.1
    # via stevedore
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via
    #   -r requirements-dev.in
    #   autoflake
    #   flake8
pylint-django==2.4.3
    # via -r requirements-dev.in
pylint-plugin-utils==0.6
    # via pylint-django
pylint==2.7.4
    # via
    #   -r requirements-dev.in
    #   pylint-django
    #   pylint-plugin-utils
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements-dev.in
pytest-django==4.2.0
    # via -r requirements-dev.in
pytest-mock==3.5.1
    # via -r requirements-dev.in
pytest==6.2.3
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
python-dateutil==2.8.1
    # via freezegun
pytz==2021.1
    # via
    #   -c requirements.txt
    #   -r requirements-dev.in
pyyaml==5.4.1
    # via bandit
regex==2021.4.4
    # via black
requests-mock==1.9.1
    # via -r requirements-dev.in
requests==2.25.1
    # via
    #   -c requirements.txt
    #   docker
    #   requests-mock
rope==0.19.0
    # via -r requirements-dev.in
six==1.15.0
    # via
    #   -c requirements.txt
    #   bandit
    #   python-dateutil
    #   requests-mock
    #   websocket-client
smmap==4.0.0
    # via gitdb
sqlalchemy==1.4.10
    # via -r requirements-dev.in
stevedore==3.3.0
    # via bandit
tblib==1.7.0
    # via -r requirements-dev.in
testcontainers==3.4.0
    # via -r requirements-dev.in
toml==0.10.2
    # via
    #   -c requirements.txt
    #   black
    #   pylint
    #   pytest
typed-ast==1.4.3
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   gitpython
    #   mypy
urllib3==1.26.4
    # via
    #   -c requirements.txt
    #   requests
websocket-client==0.58.0
    # via docker
wrapt==1.12.1
    # via
    #   astroid
    #   testcontainers
